requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-bootstrap-pagination==1.7.1
django-crispy==1.8.1
django-crispy-forms==1.9.0
et-xmlfile==1.0.1
fpdf==1.7.2
ghostscript==0.6
idna==2.9
jdcal==1.4.1
kombu==4.6.8
oauthlib==3.1.0
openpyxl==3.0.3
order==1.2.0
pdf2image==1.12.1
Pillow==7.0.0
PyMuPDF==1.16.17
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
scinum==1.0.2
six==1.14.0
sqlparse==0.3.1
stripe==2.46.0
urllib3==1.25.8
vine==1.3.0
